requirements.txt
aiohttp
asyncio
discord.py
reactionmenu
discord.py-pagination